moved main outside of utils (abaut time)
